module rename and restructure
