+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/scheduler/_v2/__init__.py
@@ -1,7 +1,8 @@
 from ._action import Action
 from ._marker import mark_step
-from ._workflow import Workflow, WorkflowRunnerManager
+from ._workflow import Workflow
 from ._workflow_context import WorkflowContext
+from ._workflow_runner_manager import WorkflowRunnerManager
 
 __all__: tuple[str, ...] = (
     "mark_step",
